% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalviz-package.R
\docType{package}
\name{gdalviz-package}
\alias{gdalviz}
\alias{gdalviz-package}
\title{gdalviz: Visualize Modern GDAL Pipelines}
\description{
Produces interactive or static visualizations of modern GDAL CLI pipelines.
}
\author{
\strong{Maintainer}: Jimmy Briggs \email{jimmy.briggs@jimbrig.com} (\href{https://orcid.org/0000-0002-7489-8787}{ORCID})

Authors:
\itemize{
  \item Jimmy Briggs \email{jimmy.briggs@jimbrig.com} (\href{https://orcid.org/0000-0002-7489-8787}{ORCID})
}

}
\keyword{internal}
