% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-reactflow.R
\name{pipelineFlowOutput}
\alias{pipelineFlowOutput}
\alias{renderPipelineFlow}
\title{Shiny bindings for pipeline_flow widgets}
\usage{
pipelineFlowOutput(outputId, width = "100\%", height = "560px")

renderPipelineFlow(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{Output variable to read from.}

\item{width, height}{CSS dimensions for the widget container.}

\item{expr}{An expression that returns a \code{pipeline_flow} widget.}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Whether \code{expr} is a quoted expression.}
}
\value{
\code{pipelineFlowOutput()} returns a Shiny output element;
\code{renderPipelineFlow()} returns a Shiny render function.
}
\description{
Output and render functions for using \code{\link[=render_reactflow]{render_reactflow()}} widgets within
Shiny applications.
}
