% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_gdalg}
\alias{read_gdalg}
\title{Read a GDAL pipeline from a GDALG file or raw string}
\usage{
read_gdalg(x, contract = gdalviz_contract())
}
\arguments{
\item{x}{A path to a GDALG/JSON/text file, or a raw pipeline string.}

\item{contract}{A \code{gdalviz_contract}.}
}
\value{
A \code{gdalviz_pipeline}.
}
\description{
Accepts a path to a GDALG JSON file (\verb{\{"type":"gdal_streamed_alg", "command_line": "..."\}}), a path to a raw pipeline text file, or a raw
pipeline string, and returns a parsed \code{gdalviz_pipeline}.
}
