% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{read_script}
\alias{read_script}
\title{Read a shell or powershell script file into a pipeline}
\usage{
read_script(path, shell = NULL, contract = gdalviz_contract())
}
\arguments{
\item{path}{Path to a \code{.sh} / \code{.ps1} script.}

\item{shell}{Source shell: \code{"bash"} (default) or \code{"powershell"}.}

\item{contract}{A \code{gdalviz_contract}.}
}
\value{
A \code{gdalviz_pipeline}.
}
\description{
Read a shell or powershell script file into a pipeline
}
