% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalviz-package.R
\docType{package}
\name{gdalviz-package}
\alias{gdalviz}
\alias{gdalviz-package}
\title{gdalviz: Visualize Modern GDAL Pipelines}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Produces interactive or static visualizations of modern GDAL CLI pipelines.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://docs.jimbrig.com/gdalviz/}
  \item \url{https://github.com/jimbrig/gdalviz}
  \item Report bugs at \url{https://github.com/jimbrig/gdalviz/issues}
}

}
\author{
\strong{Maintainer}: Jimmy Briggs \email{jimmy.briggs@jimbrig.com} (\href{https://orcid.org/0000-0002-7489-8787}{ORCID})

Authors:
\itemize{
  \item Jimmy Briggs \email{jimmy.briggs@jimbrig.com} (\href{https://orcid.org/0000-0002-7489-8787}{ORCID})
}

}
\keyword{internal}
